simplify: Extract code from glview
NFC
